Move GetProteinTestCase to test_input
